split multiform
