+++ src/multiform/internal/index.ts
@@ -1,0 +1,55 @@
+/**
+ * Created by sam on 26.12.2016.
+ */
+
+
+import {IVisMetaData, IVisPluginDesc} from '../../vis';
+
+export class ProxyMetaData implements IVisMetaData {
+  constructor(private proxy: () => IVisMetaData) {
+
+  }
+
+  get scaling() {
+    const p = this.proxy();
+    return p ? p.scaling : 'free';
+  }
+
+  get rotation() {
+    const p = this.proxy();
+    return p ? p.rotation : 0;
+  }
+
+  get sizeDependsOnDataDimension() {
+    const p = this.proxy();
+    return p ? p.sizeDependsOnDataDimension : [false, false];
+  }
+}
+
+export function selectVis(initial: number|string|IVisPluginDesc, visses: IVisPluginDesc[]) {
+  switch (typeof initial) {
+    case 'number':
+      return visses[Math.max(0, Math.min(<number>initial, visses.length - 1))];
+    case 'string':
+      return visses[Math.max(0, visses.findIndex((v) => v.id === <string>initial))];
+    default:
+      return visses[Math.max(0, visses.indexOf(<IVisPluginDesc>initial))];
+  }
+}
+
+
+export function clearNode(parent: Element) {
+  let node = parent.firstChild;
+  while ((node = parent.firstChild) != null) {
+    parent.removeChild(node);
+  }
+}
+
+export function createNode(parent: HTMLElement, type: string = 'div', clazz?: string) {
+  let node = parent.ownerDocument.createElement(type);
+  if (clazz) {
+    clazz.split(' ').forEach((c) => node.classList.add(c));
+  }
+  parent.appendChild(node);
+  return node;
+}
